+++ code/src/cljs/sixsq/nuvla/ui/deployments_detail/views.cljs
@@ -14,6 +14,7 @@
             [sixsq.nuvla.ui.deployments-detail.subs :as subs]
             [sixsq.nuvla.ui.deployments.subs :as deployments-subs]
             [sixsq.nuvla.ui.deployments.utils :as deployments-utils]
+            [sixsq.nuvla.ui.deployments.utils :as utils]
             [sixsq.nuvla.ui.i18n.subs :as i18n-subs]
             [sixsq.nuvla.ui.job.subs :as job-subs]
             [sixsq.nuvla.ui.job.views :as job-views]
@@ -70,7 +71,7 @@
         parameters (subscribe [::subs/deployment-parameters])
         running    (sum-running-replicas @parameters)
         desired    (sum-desired-replicas @parameters)]
-    (when (and (= state "STARTED") running desired (not= running desired))
+    (when (and (= state utils/STARTED) running desired (not= running desired))
       [ui/Segment
        [ui/Progress {:label    "deployment: started (replicas: running/required)"
                      :total    desired
@@ -460,7 +461,7 @@
 (defn StartUpdateButton
   [{:keys [data state] :as deployment}]
   (let [tr         (subscribe [::i18n-subs/tr])
-        start      (#{"CREATED" "STOPPED"} state)
+        start      (#{"CREATED" utils/STOPPED} state)
         first-step (if data :data :infra-services)
         button     (action-button
                      {:button-text (if start
@@ -773,7 +774,7 @@
 (defn depl-state->status
   [state]
   (case (if (some? state) (str/lower-case state) "")
-    "started" :online
+    utils/STARTED :online
     :offline))
 
 
@@ -810,7 +811,9 @@
              :on             "hover"
              :size           "tiny"
              :hide-on-scroll true}] ": "]
-          state]]))))
+          (if (= state utils/STARTED)
+            utils/RUNNING
+            state)]]))))
 
 
 (defn DeploymentDetails